refactor: centralize project metadata utilities
